+++ src/main/java/xyz/roosterseatyou/marvelitems/api/events/EventHandler.java
@@ -9,7 +9,7 @@ import org.bukkit.event.inventory.InventoryClickEvent;
 import org.bukkit.inventory.InventoryView;
 import org.bukkit.inventory.PlayerInventory;
 import xyz.roosterseatyou.marvelitems.api.enums.StoneType;
-import xyz.roosterseatyou.marvelitems.events.infinity.GauntletEvents;
+import xyz.roosterseatyou.marvelitems.events.items.infinity.GauntletEvents;
 import xyz.roosterseatyou.marvelitems.utils.MarvelUtils;
 
 import java.util.UUID;
